chore: rename MonthNavigator to month-navigator to follow naming conventions
